move all tests to unity project
