+++ unity/Assets/Editor/CommonRequests.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4a6fbc2005e123748ba3feb321973b19
+timeCreated: 1467401363
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
