+++ unity/Assets/Editor/SignatureBuilderTests.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1a896986efcbea8478dae4a5f68f847c
+timeCreated: 1467400516
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
