+++ unity/Assets/Services.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 259c5789c50235f4fa40eece84790c22
+folderAsset: yes
+timeCreated: 1467397936
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
