+++ unity/Assets/Services/ScoreboardModels.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 859a13ca9b5eb6b48b43dc7832fcd161
+timeCreated: 1467397943
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
