+++ unity/Assets/Utils.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 3c763fca06bb94f4389206d5b9aaa090
+folderAsset: yes
+timeCreated: 1467397937
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
