+++ unity/Assets/Utils/SignatureBuilder.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1d0d51f2b6613c54da36a261168caa55
+timeCreated: 1467397937
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
